+++ app/src/main/java/nextstep/signup/ui/component/SignUpButton.kt
@@ -1,0 +1,40 @@
+package nextstep.signup.ui.component
+
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import nextstep.signup.R
+import nextstep.signup.ui.theme.Blue50
+import nextstep.signup.ui.theme.RobotoMedium
+
+@Composable
+fun SignUpButton() {
+    Button(
+        onClick = {},
+        colors = ButtonDefaults.buttonColors(containerColor = Blue50),
+        contentPadding = PaddingValues(vertical = 15.dp),
+        modifier = Modifier.fillMaxWidth()
+    ) {
+        Text(
+            text = stringResource(R.string.signup_button),
+            fontSize = 14.sp,
+            color = Color.White,
+            fontFamily = RobotoMedium,
+        )
+    }
+}
+
+@Preview
+@Composable
+private fun SignUpButtonPreview() {
+    SignUpButton()
+}
